% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_by_index.R
\name{get_color_by_index}
\alias{get_color_by_index}
\title{Return the color corresponding to an integer index value, with redish negative values and blueish positive values.}
\usage{
get_color_by_index(integer_index)
}
\arguments{
\item{integer_index}{The index value of the value that is to be retrieved.}
}
\value{
The color in hex RGB format.
}
\description{
`get_color_by_index(integer_index)` return the color corresponding to the index value.
}
\examples{
get_color_by_index(-1);

\dontrun{
get_color_by_index(-1);
}

}
