% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamperf2020data.R
\docType{data}
\name{iamperf2020data}
\alias{iamperf2020data}
\title{The data of the 2020 IAM Performance Measurement Survey}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.open-measure.org}{Open-Measure}
}
\usage{
data(iamperf2020data)
}
\description{
This data contains the survey data after the following transformation steps:
1). Extraction from QuestionPro in SPSS format.
2). Opening and re-saving in SPSS format with IBM SPSS to solve a technical issue with the file exported from QuestionPro.
3). Data importation in R using the setup_data package function.
4). Extraction of the dataset dimensions that do not contain PII information.
5). Saving of the data in .RData format in the package /data folder.
}
\examples{
data(iamperf2020data);
head(iamperf2020data);
}
\references{
The Open-Measure Project
(\href{https://www.open-measure.org}{Open-Measure})
}
\keyword{datasets}
