% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dataset.R
\name{setup_dataset}
\alias{setup_dataset}
\title{Prepare an anonimized version of the IAM Performance Measurement 2020 Survey
for usage as part of the iamperf2020 R package.}
\usage{
setup_dataset()
}
\value{
nothing.
}
\description{
`setup_dataset()` Prepare the dataset and output the package RData file in the /data folder.
}
\examples{

\dontrun{
setup_dataset();
}

}
