% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q18_organization_size.R
\name{visualize_q18_organization_size}
\alias{visualize_q18_organization_size}
\title{Visualize question #18: organization size.}
\usage{
visualize_q18_organization_size()
}
\value{
plot output.
}
\description{
`visualize_organization_size()` visualizes question data.
}
\examples{

\dontrun{
visualize_q18_organization_size();
}

}
