% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q20_dedicated_team.R
\name{visualize_q20_dedicated_team}
\alias{visualize_q20_dedicated_team}
\title{Visualize question #20: shared versus dedicated team.
References:
- https://cran.r-project.org/web/packages/likert/likert.pdf
- https://rcompanion.org/handbook/E_03.html
- https://rdrr.io/github/decisionpatterns/stringr.tools/
- https://rdrr.io/github/decisionpatterns/stringr.tools/man/str_prefix.html}
\usage{
visualize_q20_dedicated_team()
}
\value{
plot output.
}
\description{
`visualize_q20_dedicated_team()` visualizes question data.
}
\examples{
visualize_q20_dedicated_team();

\dontrun{
visualize_q20_dedicated_team();
}

}
