% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q21_organization_model.R
\name{visualize_q21_organization_model}
\alias{visualize_q21_organization_model}
\title{Visualize question #20: shared versus dedicated team.
References:
- https://cran.r-project.org/web/packages/likert/likert.pdf
- https://rcompanion.org/handbook/E_03.html
- https://rdrr.io/github/decisionpatterns/stringr.tools/
- https://rdrr.io/github/decisionpatterns/stringr.tools/man/str_prefix.html}
\usage{
visualize_q21_organization_model()
}
\value{
plot output.
}
\description{
`visualize_q21_organization_model()` visualizes question data.
}
\examples{
visualize_q21_organization_model();

\dontrun{
visualize_q21_organization_model();
}

}
