% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q22_iam_manager_reporting_line.R
\name{visualize_q22_iam_manager_reporting_line}
\alias{visualize_q22_iam_manager_reporting_line}
\title{Visualize question #22: IAM manager reporting line
References:
- https://cran.r-project.org/web/packages/eulerr/vignettes/visualization.html
- https://stackoverflow.com/questions/46044564/customizing-euler-diagram-colors-with-eulerr-r}
\usage{
visualize_q22_iam_manager_reporting_line()
}
\value{
plot output.
}
\description{
`visualize_q22_iam_manager_reporting_line()` visualizes question data.
}
\examples{
visualize_q22_iam_manager_reporting_line();

\dontrun{
visualize_q22_iam_manager_reporting_line();
}

}
