% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q23_iam_goals.R
\name{visualize_q23_iam_goals}
\alias{visualize_q23_iam_goals}
\title{Visualize question #23: IAM goals
References:
- https://cran.r-project.org/web/packages/likert/likert.pdf
- https://rcompanion.org/handbook/E_03.html
- https://rdrr.io/github/decisionpatterns/stringr.tools/
- https://rdrr.io/github/decisionpatterns/stringr.tools/man/str_prefix.html}
\usage{
visualize_q23_iam_goals()
}
\value{
plot output.
}
\description{
`visualize_q23_iam_goals()` visualizes question data.
}
\examples{
visualize_q23_iam_goals();

\dontrun{
visualize_q23_iam_goals();
}

}
