% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q24_maturity_level.R
\name{visualize_q24_maturity_level}
\alias{visualize_q24_maturity_level}
\title{Visualize question #24: maturity level.
References:
- https://cran.r-project.org/web/packages/likert/likert.pdf
- https://rcompanion.org/handbook/E_03.html
- https://rdrr.io/github/decisionpatterns/stringr.tools/
- https://rdrr.io/github/decisionpatterns/stringr.tools/man/str_prefix.html}
\usage{
visualize_q24_maturity_level()
}
\value{
plot output.
}
\description{
`visualize_q24_maturity_level()` visualizes question data.
}
\examples{
visualize_q24_maturity_level();

\dontrun{
visualize_q24_maturity_level();
}

}
