% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q27_indicator_design.R
\name{visualize_q27_indicator_design}
\alias{visualize_q27_indicator_design}
\title{Visualize question #27: indicator design.
References:
- https://cran.r-project.org/web/packages/likert/likert.pdf
- https://rcompanion.org/handbook/E_03.html
- https://rdrr.io/github/decisionpatterns/stringr.tools/
- https://rdrr.io/github/decisionpatterns/stringr.tools/man/str_prefix.html}
\usage{
visualize_q27_indicator_design()
}
\value{
plot output.
}
\description{
`visualize_q27_indicator_design()` visualizes question data.
}
\examples{
visualize_q27_indicator_design();

\dontrun{
visualize_q27_indicator_design();
}

}
