% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q28_indicator_framework.R
\name{visualize_q28_indicator_framework}
\alias{visualize_q28_indicator_framework}
\title{Visualize question #28: indicator framework.
References:
- https://cran.r-project.org/web/packages/likert/likert.pdf
- https://rcompanion.org/handbook/E_03.html
- https://rdrr.io/github/decisionpatterns/stringr.tools/
- https://rdrr.io/github/decisionpatterns/stringr.tools/man/str_prefix.html}
\usage{
visualize_q28_indicator_framework()
}
\value{
plot output.
}
\description{
`visualize_q28_indicator_framework()` visualizes question data.
}
\examples{
visualize_q28_indicator_framework();

\dontrun{
visualize_q28_indicator_framework();
}

}
