% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q29_indicator_communication.R
\name{visualize_q29_indicator_communication}
\alias{visualize_q29_indicator_communication}
\title{Visualize question #29: indicator communication.
References:
- https://cran.r-project.org/web/packages/likert/likert.pdf
- https://rcompanion.org/handbook/E_03.html
- https://rdrr.io/github/decisionpatterns/stringr.tools/
- https://rdrr.io/github/decisionpatterns/stringr.tools/man/str_prefix.html}
\usage{
visualize_q29_indicator_communication()
}
\value{
plot output.
}
\description{
`visualize_q29_indicator_communication()` visualizes question data.
}
\examples{
visualize_q29_indicator_communication();

\dontrun{
visualize_q29_indicator_communication();
}

}
