% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q30_indicator_automation.R
\name{visualize_q30_indicator_automation}
\alias{visualize_q30_indicator_automation}
\title{Visualize question #30: indicator automation.}
\usage{
visualize_q30_indicator_automation(
  legend = TRUE,
  sub_title = TRUE,
  title = TRUE
)
}
\arguments{
\item{legend}{(boolean) Whether the graph legend should be plotted.}

\item{sub_title}{(boolean) Whether the graph sub title should be plotted.}

\item{title}{(boolean) Whether the graph title should be plotted.}
}
\value{
plot output.
}
\description{
`visualize_q29_indicator_communication()` visualizes question data.
}
\examples{
visualize_q30_indicator_automation();
visualize_q30_indicator_automation(legend = FALSE, sub_title = FALSE, title = FALSE);

\dontrun{
visualize_q30_indicator_automation();
}

}
