% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q31_indicator_coverage.R
\name{visualize_q31_indicator_coverage}
\alias{visualize_q31_indicator_coverage}
\title{Visualize question #31: indicator coverage
References:
- https://cran.r-project.org/web/packages/likert/likert.pdf
- https://rcompanion.org/handbook/E_03.html
- https://rdrr.io/github/decisionpatterns/stringr.tools/
- https://rdrr.io/github/decisionpatterns/stringr.tools/man/str_prefix.html}
\usage{
visualize_q31_indicator_coverage()
}
\value{
plot output.
}
\description{
`visualize_q31_indicator_coverage()` visualizes question data.
}
\examples{
visualize_q31_indicator_coverage();

\dontrun{
visualize_q31_indicator_coverage();
}

}
