% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spssfile_load.R
\name{spssfile_load}
\alias{spssfile_load}
\title{Wrapper function to load SPSS files.}
\usage{
spssfile_load(spssfile_path = NULL)
}
\arguments{
\item{spssfile_path:}{the path of the SPSS file.}
}
\value{
output the SPSS file data frame.
}
\description{
References:
- http://www.milanor.net/blog/how-to-open-an-spss-file-into-r/
- https://stat.ethz.ch/R-manual/R-devel/library/foreign/html/read.spss.html
}
\details{
`spssfile_load(spssfile_path)` load the raw survey data in SPSS format.
}
\examples{
spssfile_load(spssfile_path);

\dontrun{
spssfile_load(spssfile_path);
}

}
