% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_q30_indicator_automation.R
\name{visualize_q30_indicator_automation}
\alias{visualize_q30_indicator_automation}
\title{Visualize question #30: indicator automation.}
\usage{
visualize_q30_indicator_automation()
}
\value{
plot output.
}
\description{
`visualize_q29_indicator_communication(survey_rawdata)` visualizes question data.
}
\examples{
visualize_q30_indicator_automation();

\dontrun{
visualize_q30_indicator_automation();
}

}
